Initial implementation of custom RGB lighting
This is a checkpoint commit. I think that I have things in a working state. Frustratingly, my keyboard has kept crashing today as I've been iterating on this functionality. My best guess at this point is that it has something to do with the brightness of the RGB LEDs. Bitshifting the colors down one seems to have stopped the crashing, but it's nondeterministic and can happen even after a long period of time, so I can't even be sure of that at this point. Committing this as what I think is a working checkpoint, and will pick this back up to improve the implementation later.