Use total current-aware scaling to ensure that total RGB LED brightness does not exceed a given threshold
I am confident that the crashes that I have been intermittently experiencing have to do with the lighting code, and I'm pretty sure that they have to do with the total power draw from the lighting. This commit attempts to scale the overall sum of light output down to the given percentage of the maximum possible light output. I will play around with the value for a bit and test out different budgets to see what the maximum safe value seems to be.