Move lock screen double tap to boot key
The tap dance on Q was making caps word end on Q, so I couldn't type SQL with caps word. I know that the correct solution here is to add custom handling for caps word ending so this keycode doesn't end it, but I also was having an issue where there are actually times where I want to type qq, which was making my computer lock if I did it too quickly. So this just feels like a better solution overall. I was surprised that I didn't end up needing to use the tap dance logic aliases for this functionality.