Require tap for combos containing any keys with hold functionality
I haven't been able to easily shift+command+click anymore because F+D is shift+command but is also the enter combo. This logic makes it so that F+D will only be interpreted as the enter combo if the combo is tapped, not held. The logic is generic and applies to all combos containing keys with hold functionality.